requirements.txt
# Discord API
discord.py[voice]>=1.5.1

# Youtube API
google-api-python-client>=1.12.5
youtube_dl>=2020.11.12